bug fixed in xdelta_hzr packer  - bugfix: bytes to encode are recursively increased if data restoration check fails  - various packers tried out including lz4 and zzlib  - some warnings removed

+++ lib_rspt/lib_signalpacker/signal_packer_hadamard.cpp
@@ -68,7 +68,7 @@ public:
             for (int i = 0; i < enc_.d2; ++i)
                 enc_.d2d[j][i] = hadamard.d2d[j][i];
         }
-        unsigned char header[enc_.d1 * 3];
+        unsigned char header[(uint16_t)enc_.d1 * 3];
         for (int i = 0; i < enc_.d1; ++i)
         {
             header[i * 3 + 0] = means[i] & 0x000000FF;
